GitWrap/Errors: Rename Git_p.h to GitWrapPrivate.h
