+++ src/Libs/GitWrap/GitWrapPrivate.h
