+++ src/Libs/GitWrap/ObjectBlob.cpp
@@ -14,7 +14,7 @@
  *
  */
 
-#include "Git_p.h"
+#include "GitWrapPrivate.h"
 
 #include "ObjectBlob.h"
