+++ src/Libs/GitWrap/ObjectCommit.cpp
@@ -14,7 +14,7 @@
  *
  */
 
-#include "Git_p.h"
+#include "GitWrapPrivate.h"
 
 #include "ObjectPrivate.h"
 #include "ObjectCommit.h"
