+++ src/Libs/GitWrap/ObjectTree.cpp
@@ -14,7 +14,7 @@
  *
  */
 
-#include "Git_p.h"
+#include "GitWrapPrivate.h"
 
 #include "DiffList.h"
 #include "DiffListPrivate.h"
