+++ src/Libs/GitWrap/Signature.cpp
@@ -14,7 +14,7 @@
  *
  */
 
-#include "Git_p.h"
+#include "GitWrapPrivate.h"
 #include "Signature.h"
 
 namespace Git
